Add 'stamp' command for updating the .stamp file in data/ folders.
You run this occasionally to mark the folders as "a computer is at least
watching this folder."  Then if a new log message doesn't show up, you know
something weird is going on (and the 'expire' script will log about it).

You can use this as a heartbeat indicator for remote computers.